+++ src/components/home.js
@@ -1,5 +1,6 @@
 import React, { useContext, useEffect, useRef } from 'react';
 import { ThemeContext } from '../master/theme';
+import { ContentCategory } from '../master/switchcategory';
 import Glide from '@glidejs/glide';
 import '@glidejs/glide/dist/css/glide.core.min.css';
 
@@ -76,8 +77,10 @@ const Home = ({ openSwitch }) => {
 
   const images = imageMap[theme] || [];
 
+  const { category } = useContext(ContentCategory);
+
   return (
-    <div className="min-h-screen w-full flex flex-col">
+    <div className={`${category === 'home' ? 'flex flex-col min-h-screen w-full' : 'hidden' }`}>
       <div className={`w-full pt-12 bg-${theme} bg-opacity-50`}>
         <div className="glide">
           <div className="glide__track" data-glide-el="track">
@@ -98,7 +101,7 @@ const Home = ({ openSwitch }) => {
       <div className="flex-1 w-full px-2 sm:px-8 py-4 flex bg-cover bg-bottom bg-no-repeat saturate-150 relative" style={{ backgroundImage: `url(${background1})` }}>
         <div className={`absolute inset-0 bg-${theme} bg-opacity-60 z-0`} />
           <div className={`relative w-full flex flex-col justify-center items-center p-4 border-t-8 bg-slate-100 bg-opacity-60 border-${theme} rounded-lg shadow-2xl`}>
-            <h1 className='ff-goldman text-4xl md:text-5xl sm:text-4xl text-slate-950 text-center'><span className={`text-${theme}`}>-</span>{ character.name }<span className={`text-${theme}`}>-</span></h1>
+            <h1 className='ff-goldman text-4xl md:text-5xl sm:text-4xl text-slate-950 text-center overflow-y-hidden'><span className={`text-${theme}`}>-</span>{ character.name }<span className={`text-${theme}`}>-</span></h1>
             <p className={`ff-be-vietnam-pro font-bold text-center ease-in-out transition-all text-slate-950 duration-300 text-lg hover-text-${theme}`}>{character.japanese}</p>
             <p className={`ff-be-vietnam-pro font-normal text-center ease-in-out transition-all text-slate-950 duration-300 text-lg`}>CV: {character.va_japanese}</p>
             <button onClick={openSwitch} type='button' className={`h-12 w-40 bg-${theme} transition-all duration-300 ease-in-out hover:brightness-110 rounded-lg`}>