+++ src/index.js
@@ -7,11 +7,17 @@ import './App.css';
 import DefaultBG from './master/default_bg';
 // Color Pallete
 import { ThemeProvider } from './master/theme';
+// Content Category
+import { ContentProvider } from './master/switchcategory';
 
 // Components
 import LoadingScreen from './master/loadingscreen';
 import NavbarWrapper from './components/navwrapper';
 import SwitchCharacter from './master/switchcharacter';
+import Gallery from './components/gallery';
+import Profile from './components/profile';
+import Trivia from './components/trivia';
+import Credits from './components/credits';
 
 const main = ReactDOM.createRoot(document.getElementById('main'));
 main.render(
@@ -19,9 +25,15 @@ main.render(
   <React.StrictMode>
     <DefaultBG>
       <ThemeProvider>
-        <LoadingScreen />
-        <NavbarWrapper />
-        <SwitchCharacter/>
+        <ContentProvider>
+          <LoadingScreen />
+          <NavbarWrapper />
+          <SwitchCharacter/>
+          <Gallery />
+          <Profile />
+          <Trivia />
+          <Credits />
+        </ContentProvider>
       </ThemeProvider>
     </DefaultBG>
   </React.StrictMode>